requirement.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
pydantic>=2.6.0 --pre